+++ lib/internal/streams/pipeline.js
@@ -11,7 +11,6 @@ const {
   FunctionPrototypeCall,
   ReflectApply,
   SymbolAsyncIterator,
-  SymbolIterator,
 } = primordials;
 
 let eos;
@@ -26,6 +25,12 @@ const {
 } = require('internal/errors').codes;
 
 const { validateCallback } = require('internal/validators');
+
+const {
+  isIterable,
+  isReadable,
+  isStream,
+} = require('internal/streams/utils');
 
 let EE;
 let PassThrough;
@@ -80,26 +85,6 @@ function popCallback(streams) {
   // checking for length === 0 as well.
   validateCallback(streams[streams.length - 1]);
   return ArrayPrototypePop(streams);
-}
-
-function isReadable(obj) {
-  return !!(obj && typeof obj.pipe === 'function');
-}
-
-function isWritable(obj) {
-  return !!(obj && typeof obj.write === 'function');
-}
-
-function isStream(obj) {
-  return isReadable(obj) || isWritable(obj);
-}
-
-function isIterable(obj, isAsync) {
-  if (!obj) return false;
-  if (isAsync === true) return typeof obj[SymbolAsyncIterator] === 'function';
-  if (isAsync === false) return typeof obj[SymbolIterator] === 'function';
-  return typeof obj[SymbolAsyncIterator] === 'function' ||
-    typeof obj[SymbolIterator] === 'function';
 }
 
 function makeAsyncIterable(val) {
