+++ lib/stream/promises.js
@@ -3,8 +3,6 @@
 const {
   ArrayPrototypePop,
   Promise,
-  SymbolAsyncIterator,
-  SymbolIterator,
 } = primordials;
 
 const {
@@ -15,28 +13,13 @@ const {
   validateAbortSignal,
 } = require('internal/validators');
 
+const {
+  isIterable,
+  isStream,
+} = require('internal/streams/utils');
+
 let pl;
 let eos;
-
-function isReadable(obj) {
-  return !!(obj && typeof obj.pipe === 'function');
-}
-
-function isWritable(obj) {
-  return !!(obj && typeof obj.write === 'function');
-}
-
-function isStream(obj) {
-  return isReadable(obj) || isWritable(obj);
-}
-
-function isIterable(obj, isAsync) {
-  if (!obj) return false;
-  if (isAsync === true) return typeof obj[SymbolAsyncIterator] === 'function';
-  if (isAsync === false) return typeof obj[SymbolIterator] === 'function';
-  return typeof obj[SymbolAsyncIterator] === 'function' ||
-    typeof obj[SymbolIterator] === 'function';
-}
 
 function pipeline(...streams) {
   if (!pl) pl = require('internal/streams/pipeline');
